1. enable multistream memory reuse
2. add naive assign stream logic (reverse DFS)
3. add -fnum_stream to specify the number of streams
4. fix conv kernel bugs in multistream scenario
5. replace global cudnn/cublas handle with stream specific handle

+++ src/nnfusion/common/descriptor/tensor.cpp
@@ -31,7 +31,7 @@ nnfusion::descriptor::Tensor::Tensor(const nnfusion::element::Type& element_type
                                      bool is_constant,
                                      bool is_parameter,
                                      bool is_RDMA_tensor,
-                                     size_t group_id,
+                                     int group_id,
                                      size_t device_id)
     : m_element_type(element_type)
     , m_shape(pshape.is_static() ? pshape.to_shape() : nnfusion::Shape{})
@@ -56,7 +56,7 @@ nnfusion::descriptor::Tensor::Tensor(const nnfusion::element::Type& element_type
                                      bool is_constant,
                                      bool is_parameter,
                                      bool is_RDMA_tensor,
-                                     size_t group_id,
+                                     int group_id,
                                      size_t device_id)
     : m_element_type(element_type)
     , m_shape(pshape.is_static() ? pshape.to_shape() : nnfusion::Shape{})
