+++ src/nnfusion/common/descriptor/tensor.hpp
@@ -49,7 +49,7 @@ namespace nnfusion
                    bool is_constant = false,
                    bool is_parameter = false,
                    bool is_RDMA_tensor = false,
-                   size_t group_id = -1,
+                   int group_id = -1,
                    size_t device_id = 0);
 
             Tensor(const nnfusion::element::Type& element_type,
@@ -60,7 +60,7 @@ namespace nnfusion
                    bool is_constant = false,
                    bool is_parameter = false,
                    bool is_RDMA_tensor = false,
-                   size_t group_id = -1,
+                   int group_id = -1,
                    size_t device_id = 0);
 
             const std::string& get_name() const { return m_name; }
@@ -92,8 +92,8 @@ namespace nnfusion
             void set_parameter(bool value = true) { m_parameter = value; }
             void set_RDMA(bool value = true) { m_RDMA = value; }
             //The default group_id is -1, which means the tensor does not belong to any specific group.
-            void set_group_id(size_t group_id) { m_group_id = group_id; }
-            size_t get_group_id() const { return m_group_id; }
+            void set_group_id(int group_id) { m_group_id = group_id; }
+            int get_group_id() const { return m_group_id; }
             void set_device_type(DeviceType device_type) { m_device_type = device_type; }
             DeviceType get_device_type() const { return m_device_type; }
             void set_device_id(size_t device_id) { m_device_id = device_id; }
@@ -117,7 +117,7 @@ namespace nnfusion
             bool m_constant;
             bool m_parameter;
             bool m_RDMA;
-            size_t m_group_id;
+            int m_group_id;
             DeviceType m_device_type;
             size_t m_device_id;
         };
